+++ dflib/src/main/java/com/nhl/dflib/EnumSeries.java
@@ -1,0 +1,10 @@
+package com.nhl.dflib;
+
+/**
+ * Series for non-nullable categorical data.
+ *
+ * @since 0.6
+ */
+public interface EnumSeries<T extends Enum<T>> extends Series<T> {
+
+}